requirements.txt
google-api-python-client
spotipy
pandas
networkx
numpy
scipy
matplotlib
seaborn
python-dotenv